+++ ft_constructor.c
@@ -10,7 +10,8 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
+#include "includes/minirt.h"
+#include "includes/ft_math.h"
 
 t_vec	make_vec(double x, double y, double z)
 {
@@ -70,4 +71,20 @@ t_ray	make_view(t_camera *cam, double u, double v)
 					cam->left),
 				cam->orig));
 	return (ray);
+}
+
+t_camera	make_cam(void)
+{
+	t_camera	cam;
+
+	cam.orig = make_point(0, 0, 10);
+	cam.dir = make_vec(0, 0, -1);
+	cam.left = minus_vec(
+			minus_vec(cam.orig,
+					  plus_vec(
+							  divide_vec_s(make_vec(WIN_WIDTH, 0, 0), 2),
+							  divide_vec_s(make_vec(0, WIN_HEIGHT, 0), 2)
+					  )),
+			make_vec(0, 0, 100));
+	return (cam);
 }
