+++ op_vector.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_math.h"
+#include "includes/ft_math.h"
 
 double	s2_vec(t_vec vec)
 {
