+++ test.c
@@ -10,7 +10,8 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
+#include "includes/minirt.h"
+#include "includes/ft_math.h"
 
 void	print_vec(t_vec vec)
 {
@@ -51,103 +52,58 @@ int	hit_sphere(t_sphere *sp, t_ray *ray, t_rec *rec)
 	return (1);
 }
 
-t_camera	make_cam(void)
-{
-	t_camera	cam;
-
-	cam.orig = make_point(0, 0, 10);
-	cam.dir = make_vec(0, 0, -1);
-	cam.left = minus_vec(
-			minus_vec(cam.orig,
-				plus_vec(
-					divide_vec_s(make_vec(WIN_WIDTH, 0, 0), 2),
-					divide_vec_s(make_vec(0, WIN_HEIGHT, 0), 2)
-					)),
-			make_vec(0, 0, 100));
-	return (cam);
-}
-
 int get_color(t_color c)
 {
 	return (((int)(c.x * 255) << 16) + ((int)(c.y * 255) << 8) + (int)(c.z * 255));
 }
 
-void	scene_init(t_scene *scene)
+void	scene_init_for_temp(t_scene *scene)
 {
-	double ka;
+	double		ka;
+	t_object	*obj, *obj2;
+	t_light		*light, *light2;
 
 	ka = 0.1;
+	scene->rec.tmax = 99999;
+	scene->rec.tmin = 1e-6;
+	scene->rec.albedo = make_color(0.7, 0, 0);
 	scene->ambient = make_color(ka, ka, ka);
-}
+	obj = malloc(sizeof(t_object));
+	obj2 = malloc(sizeof(t_object));
+	light = malloc(sizeof(t_light));
+	light2 = malloc(sizeof(t_light));
+	scene->light = obj;
+	obj->type = LIGHT;
+	obj->element = light;
+	obj->next = obj2;
+	obj2->type = LIGHT;
+	obj2->element = light2;
+	obj2->next = NULL;
 
-void	rec_init(t_rec *rec)
-{
-	rec->tmax = 999999;
-	rec->tmin = 1e-6;
+	light->point = make_point(10, 10, -100);
+	light->bright = 1; // not used?
+	light->color = make_color(1, 1, 1);
+
+	light2->color = make_color(1, 1, 1);
+	light2->bright = 1;
+	light2->point = make_point(0, 0, -10);
 }
 
 int color_sphere(t_sphere *sp, t_ray *ray)
 {
-	t_rec	rec;
+	t_scene scene;
+	double 	t;
 
-	rec_init(&rec);
-	if (hit_sphere(sp, ray, &rec))
-		return (get_color(mul_vec_s(make_color(rec.n.x + 1, rec.n.y + 1, rec.n.z + 1), 0.5)));
+	scene_init_for_temp(&scene);
+	if (hit_sphere(sp, ray, &scene.rec))
+//		return (lighting(&scene));
+		return (get_color(mul_vec_s(make_color(scene.rec.n.x + 1, scene.rec.n.y + 1, scene.rec.n.z + 1), 0.5)));
 	else
-		return (SKYBLUE);
-}
-
-int	lighting(t_scene *scene)
-{
-	t_color		light_color;
-	t_object	*lights;
-
-	light_color = make_color(0, 0, 0);
-	lights = scene->light;
-	while (lights)
 	{
-		if (lights->type == LIGHT)
-			light_color = plus_vec(light_color, point_light_get(scene, lights->element));
-		lights = lights->next;
+		t = 0.5 * (scene.ray.dir.y + 1.0);
+//		return (get_color(plus_vec(
+//					mul_vec_s(make_color(1, 1, 1), 1.0 - t),
+//					mul_vec_s(make_color(0.5, 0.7, 1.0), t))));
+		return (SKYBLUE);
 	}
-	light_color = plus_vec(light_color, scene->ambient);
-	return (get_color(
-			min_vec(
-				mul_vec(light_color, scene->rec.albedo),
-				make_color(1, 1, 1))
-						));
-}
-
-t_color	point_light_get(t_scene *scene, t_light *light)
-{
-	t_color	diffuse;
-	t_vec	light_dir;
-	double	kd;
-	t_color	specular;
-	t_vec	view_dir;
-	t_vec	reflect_dir;
-	double	spec;
-	double	ksn;
-	double	ks;
-
-	light_dir = unit_vec(minus_vec(light->point, scene->rec.p));
-	kd = fmax(dot_vec(scene->rec.n, light_dir), 0.0);
-	diffuse = mul_vec_s(light->color, kd);
-
-	view_dir = unit_vec(mul_vec_s(scene->ray.dir, -1));
-	reflect_dir = reflect(light_dir, scene->rec.n);
-	ksn = 30;
-	ks = 0.8;
-	spec = pow(fmax(dot_vec(view_dir, reflect_dir), 0.0), ksn);
-	specular = mul_vec_s(mul_vec_s(light->color, ks), spec);
-
-	return (plus_vec(diffuse, specular));
-}
-
-t_vec	reflect(t_vec v, t_vec n)
-{
-	return (plus_vec(
-			mul_vec_s(v, -1),
-			mul_vec_s(n,2 * dot_vec(v, n))
-			));
 }
